+++ ProductChase/frmProducts.cs
@@ -32,12 +32,13 @@ namespace ProductChase
         }
         private void clean()
         {
+            txtId.Clear();
             txtProductName.Clear();
             txtBrand.Clear();
             txtCost.Clear();
             txtPrice.Clear();
             txtStock.Clear();
-            cbmCategory.Text = string.Empty;
+            
 
         }
         private void btnList_Click(object sender, EventArgs e)
@@ -57,51 +58,102 @@ namespace ProductChase
             cbmCategory.ValueMember = "CATEGORYID";
             cbmCategory.DisplayMember = "CATEGORYNAME";
             cbmCategory.DataSource = dt;
-            cbmCategory.Text = string.Empty;
             listIt();
 
         }
 
         private void btnSave_Click(object sender, EventArgs e)
         {
-            List<string> productNameList = new List<string>();
-            SqlCommand cmd1 = new SqlCommand("Select ProductName from TBLPRODUCTS", conn.conn());
-            SqlDataReader dr = cmd1.ExecuteReader();
-            while (dr.Read())
+            if (txtProductName.Text.Trim().Length == 0 || txtBrand.Text.Trim().Length == 0 || txtCost.Text.Trim().Length == 0 || txtPrice.Text.Trim().Length == 0 || txtStock.Text.Trim().Length == 0 || cbmCategory.Text == "")
             {
-                productNameList.Add(dr[0].ToString());
-            }
-            conn.conn().Close();
-            int temp = 0;
-            for (int i = 0; i < productNameList.Count; i++)
-            {
-                if (productNameList[i] == txtProductName.Text)
-                {
-                    temp++;
-                }
-            }
-
-            if (temp > 0)
-            {
-                MessageBox.Show("This product has already been ADDED. Please try to add different product", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                MessageBox.Show("Please enter Valid Informations", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
             }
             else
             {
-                SqlCommand cmd = new SqlCommand("insert into TBLPRODUCTS (PRODUCTNAME,PRODUCTBRAND,CATEGORY,PRODUCTCOST,PRODUCTPRICE,PRODUCTSTOCK) values (@p1,@p2,@p3,@p4,@p5,@p6)", conn.conn());
-                cmd.Parameters.AddWithValue("@p1", txtProductName.Text);
-                cmd.Parameters.AddWithValue("@p2", txtBrand.Text);
-                cmd.Parameters.AddWithValue("@p3", cbmCategory.SelectedValue);
-                cmd.Parameters.AddWithValue("@p4", txtCost.Text);
-                cmd.Parameters.AddWithValue("@p5", txtPrice.Text);
-                cmd.Parameters.AddWithValue("@p6", txtStock.Text);
-                cmd.ExecuteNonQuery();
-                conn.conn().Close();
 
-                MessageBox.Show("The Product has been ADDED", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                listIt();
-                clean();
+                if (txtId.Text.Trim().Length == 0)
+                {
+                    List<string> productNameList = new List<string>();
+                    SqlCommand cmd1 = new SqlCommand("Select ProductName from TBLPRODUCTS", conn.conn());
+                    SqlDataReader dr = cmd1.ExecuteReader();
+                    while (dr.Read())
+                    {
+                        productNameList.Add(dr[0].ToString());
+                    }
+                    conn.conn().Close();
+                    int temp = 0;
+                    for (int i = 0; i < productNameList.Count; i++)
+                    {
+                        if (productNameList[i] == txtProductName.Text)
+                        {
+                            temp++;
+                        }
+                    }
+
+                    if (temp > 0)
+                    {
+                        MessageBox.Show("This product has already been ADDED. Please try to add different product", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        SqlCommand cmd = new SqlCommand("insert into TBLPRODUCTS (PRODUCTNAME,PRODUCTBRAND,CATEGORY,PRODUCTCOST,PRODUCTPRICE,PRODUCTSTOCK) values (@p1,@p2,@p3,@p4,@p5,@p6)", conn.conn());
+                        cmd.Parameters.AddWithValue("@p1", txtProductName.Text);
+                        cmd.Parameters.AddWithValue("@p2", txtBrand.Text);
+                        cmd.Parameters.AddWithValue("@p3", cbmCategory.SelectedValue);
+                        cmd.Parameters.AddWithValue("@p4", txtCost.Text);
+                        cmd.Parameters.AddWithValue("@p5", txtPrice.Text);
+                        cmd.Parameters.AddWithValue("@p6", txtStock.Text);
+                        cmd.ExecuteNonQuery();
+                        conn.conn().Close();
+
+                        MessageBox.Show("The Product has been ADDED", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        listIt();
+                        clean();
+                    }
+                }
+                else
+                {
+                    List<string> productIdList = new List<string>();
+                    SqlCommand cmd1 = new SqlCommand("Select ProductId from TBLPRODUCTS", conn.conn());
+                    SqlDataReader dr = cmd1.ExecuteReader();
+                    while (dr.Read())
+                    {
+                        productIdList.Add(dr[0].ToString());
+                    }
+                    conn.conn().Close();
+                    int temp = 0;
+                    for (int i = 0; i < productIdList.Count; i++)
+                    {
+                        if (productIdList[i] == txtId.Text)
+                        {
+                            temp++;
+                        }
+                    }
+
+                    if (temp > 0)
+                    {
+                        MessageBox.Show("The Product Id has been used for another product. Please clear the form and try to save a new one", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    }
+                    else
+                    {
+                        SqlCommand cmd = new SqlCommand("insert into TBLPRODUCTS (PRODUCTNAME,PRODUCTBRAND,CATEGORY,PRODUCTCOST,PRODUCTPRICE,PRODUCTSTOCK) values (@p1,@p2,@p3,@p4,@p5,@p6)", conn.conn());
+                        cmd.Parameters.AddWithValue("@p1", txtProductName.Text);
+                        cmd.Parameters.AddWithValue("@p2", txtBrand.Text);
+                        cmd.Parameters.AddWithValue("@p3", cbmCategory.SelectedValue);
+                        cmd.Parameters.AddWithValue("@p4", txtCost.Text);
+                        cmd.Parameters.AddWithValue("@p5", txtPrice.Text);
+                        cmd.Parameters.AddWithValue("@p6", txtStock.Text);
+                        cmd.ExecuteNonQuery();
+                        conn.conn().Close();
+
+                        MessageBox.Show("The Product has been ADDED", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        listIt();
+                        clean();
+
+                    }
+
+                }
             }
-
         }
 
         private void cbmCategory_SelectedIndexChanged(object sender, EventArgs e)
@@ -113,25 +165,37 @@ namespace ProductChase
             DataTable dt = new DataTable();
             da.Fill(dt);
             dataGridView1.DataSource = dt;
-        
+
         }
 
         private void btnUpdate_Click(object sender, EventArgs e)
         {
-            SqlCommand cmd = new SqlCommand("UPDATE TBLPRODUCTS SET PRODUCTNAME=@p1,PRODUCTBRAND=@p2,CATEGORY=@p3,PRODUCTCOST=@p4,PRODUCTPRICE=@p5,PRODUCTSTOCK=@p6 where productid=@p7", conn.conn());
-            cmd.Parameters.AddWithValue("@p1", txtProductName.Text);
-            cmd.Parameters.AddWithValue("@p2", txtBrand.Text);
-            cmd.Parameters.AddWithValue("@p3", cbmCategory.SelectedValue);
-            cmd.Parameters.AddWithValue("@p4", txtCost.Text);
-            cmd.Parameters.AddWithValue("@p5", txtPrice.Text);
-            cmd.Parameters.AddWithValue("@p6", txtStock.Text);
-            cmd.Parameters.AddWithValue("@p7", txtId.Text);
-            cmd.ExecuteNonQuery();
-            conn.conn().Close();
+            if (txtProductName.Text.Trim().Length == 0 || txtBrand.Text.Trim().Length == 0 || txtCost.Text.Trim().Length == 0 || txtPrice.Text.Trim().Length == 0 || txtStock.Text.Trim().Length == 0 || cbmCategory.Text == "")
+            {
+                MessageBox.Show("Please enter Valid Informations", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+            else
+            {
+                DialogResult result = MessageBox.Show("Are you sure to UPDATE product Id: " + txtId.Text, "Information", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
 
-            MessageBox.Show("The Product has been UPDATED", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            listIt();
-            clean();
+                if (result == DialogResult.Yes)
+                {
+                    SqlCommand cmd = new SqlCommand("UPDATE TBLPRODUCTS SET PRODUCTNAME=@p1,PRODUCTBRAND=@p2,CATEGORY=@p3,PRODUCTCOST=@p4,PRODUCTPRICE=@p5,PRODUCTSTOCK=@p6 where productid=@p7", conn.conn());
+                    cmd.Parameters.AddWithValue("@p1", txtProductName.Text);
+                    cmd.Parameters.AddWithValue("@p2", txtBrand.Text);
+                    cmd.Parameters.AddWithValue("@p3", cbmCategory.SelectedValue);
+                    cmd.Parameters.AddWithValue("@p4", txtCost.Text);
+                    cmd.Parameters.AddWithValue("@p5", txtPrice.Text);
+                    cmd.Parameters.AddWithValue("@p6", txtStock.Text);
+                    cmd.Parameters.AddWithValue("@p7", txtId.Text);
+                    cmd.ExecuteNonQuery();
+                    conn.conn().Close();
+
+                    MessageBox.Show("product Id: " + txtId.Text + " has been UPDATED", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    listIt();
+                    clean();
+                }
+            }
         }
 
         private void dataGridView1_CellMouseDoubleClick(object sender, DataGridViewCellMouseEventArgs e)
@@ -143,7 +207,7 @@ namespace ProductChase
             txtId.Text = dataGridView1.Rows[choosen].Cells[0].Value.ToString();
             txtProductName.Text = dataGridView1.Rows[choosen].Cells[1].Value.ToString();
             txtBrand.Text = dataGridView1.Rows[choosen].Cells[2].Value.ToString();
-           
+
             costPriceWithCurrency = dataGridView1.Rows[choosen].Cells[4].Value.ToString();
             txtCost.Text = costPriceWithCurrency.Replace("£", "");
 
@@ -157,14 +221,26 @@ namespace ProductChase
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            SqlCommand cmd = new SqlCommand("Delete from TBLPRODUCTS where productId=@p1", conn.conn());
-            cmd.Parameters.AddWithValue("@p1",txtId.Text);
-            cmd.ExecuteNonQuery();
-            conn.conn().Close();
+            if (txtId.Text.Trim().Length == 0)
+            {
+                MessageBox.Show("Please enter a Valid Product Id by selection from table", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+            else
+            {
+                DialogResult result = MessageBox.Show("Are you sure to DELETE product Id: " + txtId.Text, "Information", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
 
-            MessageBox.Show("Product has been DELETED", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            clean();
-            listIt();
+                if (result == DialogResult.Yes)
+                {
+                    SqlCommand cmd = new SqlCommand("Delete from TBLPRODUCTS where productId=@p1", conn.conn());
+                    cmd.Parameters.AddWithValue("@p1", txtId.Text);
+                    cmd.ExecuteNonQuery();
+                    conn.conn().Close();
+
+                    MessageBox.Show("Product Id: " + txtId.Text + " has been DELETED", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    clean();
+                    listIt();
+                }
+            }
         }
     }
 }
